requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
black==24.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.4
dill==0.3.8
exceptiongroup==1.2.0
fastapi==0.109.1
fastembed==0.1.3
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langsmith==0.0.86
MarkupSafe==2.1.5
marshmallow==3.20.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.1
npy-append-array==0.9.16
numpy==1.24.4
onnx==1.15.0
onnxruntime==1.17.0
p-tqdm==1.4.0
packaging==23.2
pathos==0.3.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
pox==0.3.4
ppft==1.7.6.8
protobuf==4.25.2
pydantic==2.6.0
pydantic-core==2.16.1
PyYAML==6.0.1
qdrant-client==1.7.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.1
tomli==2.0.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
typing-extensions==4.9.0
typing-inspect==0.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
yarl==1.9.4
